Make proof assistant list on webpage (#34)
* Make proof assistant list on webpage

On index.html just list the three assistants which are apparently the
most popular.  On other.html make a new section which lists all six.
Most of the text is taken from the existing web pages, with relatively
minor tweaks to some of the language.  There is an attempt to reformat
the entries to match the format of the other tools on other.html .

* Update mmt1 entry with text from Marlo Bruder